+++ tests/Foundation/FoundationTest.php
@@ -81,7 +81,15 @@ class FoundationTest extends \PHPUnit_Framework_TestCase
             ->shouldReceive('add->title->link')->once()->andReturnNull();
         $translator->shouldReceive('get')->andReturn('foo');
         $event->shouldReceive('listen')->once()
-                ->with('orchestra.ready: admin', 'Orchestra\Foundation\AdminMenuHandler')->andReturnNull()
+                ->with('orchestra.ready: admin', 'Orchestra\Foundation\Http\Handlers\UserMenuHandler')->andReturnNull()
+            ->shouldReceive('listen')->once()
+                ->with('orchestra.ready: admin', 'Orchestra\Foundation\Http\Handlers\ExtensionMenuHandler')->andReturnNull()
+            ->shouldReceive('listen')->once()
+            ->with('orchestra.ready: admin', 'Orchestra\Foundation\Http\Handlers\SettingMenuHandler')->andReturnNull()
+            ->shouldReceive('listen')->once()
+            ->with('orchestra.ready: admin', 'Orchestra\Foundation\Http\Handlers\ResourcesMenuHandler')->andReturnNull()
+            ->shouldReceive('listen')->once()
+            ->with('orchestra.ready: admin', 'Orchestra\Foundation\AdminMenuHandler')->andReturnNull()
             ->shouldReceive('fire')->once()->with('orchestra.started', array($memoryProvider))->andReturnNull();
         $config->shouldReceive('get')->once()->with('orchestra/foundation::handles', '/')->andReturn('admin');
         $request->shouldReceive('root')->andReturn('http://localhost')